finish query for inventory , refactored models
